fix: fix code for auth google
