+++ app/api/auth/google/callback/route.js
@@ -2,16 +2,31 @@ import { NextResponse } from 'next/server'
 
 import { createClient } from '@/app/supabase/server'
 
-export async function GET() {
-  const supabase = createClient()
+// The client you created from the Server-Side Auth instructions
 
-  const { error } = await supabase.auth.getSessionFromUrl({
-    storeSession: true,
-  })
+export async function GET(request) {
+  const { searchParams, origin } = new URL(request.url)
+  const code = searchParams.get('code')
+  // if "next" is in param, use it as the redirect URL
+  const next = searchParams.get('next') ?? '/private'
 
-  if (error) {
-    return NextResponse.json({ error: error.message }, { status: 500 })
+  if (code) {
+    const supabase = createClient()
+    const { error } = await supabase.auth.exchangeCodeForSession(code)
+    if (!error) {
+      const forwardedHost = request.headers.get('x-forwarded-host') // original origin before load balancer
+      const isLocalEnv = process.env.NODE_ENV === 'development'
+      if (isLocalEnv) {
+        // we can be sure that there is no load balancer in between, so no need to watch for X-Forwarded-Host
+        return NextResponse.redirect(`${origin}${next}`)
+      } else if (forwardedHost) {
+        return NextResponse.redirect(`https://${forwardedHost}${next}`)
+      } else {
+        return NextResponse.redirect(`${origin}${next}`)
+      }
+    }
   }
 
-  return NextResponse.redirect('/private')
+  // return the user to an error page with instructions
+  return NextResponse.redirect(`${origin}/auth/auth-code-error`)
 }
